feat: Support subcommand
