+++ Sources/nurse/main.swift
@@ -1,4 +1,4 @@
 
 if #available(OSX 10.15, *) {
-    Nurse.main()
+    NurseApp.main()
 }
